Added "rebase before push" functionality to delivery check. Also added some minor improvements:     *) corrected warnings in CrazyCubeSolver_test.cpp     *) delivery check now remembers current build files

+++ Test/src/CrazyCubeSolver_test.cpp
@@ -154,9 +154,9 @@ TEST_F(CrazyCubeSolverTest, CrazyCubeSolveTest)
 		cout << "Nie umiem tego rozwiazac. Sory." << endl;
 	}
 
-	for (int i = 0; i < rozwiazania.size(); ++i)
+	for (unsigned short int i = 0; i < rozwiazania.size(); ++i)
 	{
-		for (int j = 0; j < rozwiazania[i].size(); ++j)
+		for (unsigned short int j = 0; j < rozwiazania[i].size(); ++j)
 		{
 			cout << rozwiazania[i][j] << " ";
 		}
